chore: bump dependencies versions
